refactor(web): centralize active transcript mutations
